+++ SwiftMessages/UIViewController+Extensions.swift
@@ -1,5 +1,5 @@
 //
-//  UIViewController+Utils.swift
+//  UIViewController+Extensions.swift
 //  SwiftMessages
 //
 //  Created by Timothy Moose on 8/5/16.
@@ -19,7 +19,7 @@ extension UIViewController {
         } else if case .bottom? = topBottomStyle, let tabBarController = sm_selectTabBarControllerTopDown() {
             return tabBarController
         }
-        return WindowViewController.newInstance(windowLevel: self.view.window?.windowLevel, config: config)
+        return WindowViewController.newInstance(config: config)
     }
     
     fileprivate func sm_selectNavigationControllerTopDown() -> UINavigationController? {
@@ -71,7 +71,7 @@ extension UIViewController {
             if let parent = self.parent {
                 return parent.sm_selectPresentationContextBottomUp(config)
             } else {
-                return WindowViewController.newInstance(windowLevel: self.view.window?.windowLevel, config: config)
+                return WindowViewController.newInstance(config: config)
             }
         }
         return self