+++ SwiftMessages/WindowViewController.swift
@@ -13,9 +13,12 @@ open class WindowViewController: UIViewController
     override open var shouldAutorotate: Bool {
         return config.shouldAutorotate
     }
-    
-    public init(windowLevel: UIWindow.Level?, config: SwiftMessages.Config) {
-        self.windowLevel = windowLevel ?? UIWindow.Level.normal
+
+    convenience public init() {
+        self.init(config: SwiftMessages.Config())
+    }
+
+    public init(config: SwiftMessages.Config) {
         self.config = config
         let view = PassthroughView()
         let window = PassthroughWindow(hitTestView: view)
@@ -23,25 +26,28 @@ open class WindowViewController: UIViewController
         super.init(nibName: nil, bundle: nil)
         self.view = view
         window.rootViewController = self
-        window.windowLevel = windowLevel ?? UIWindow.Level.normal
+        window.windowLevel = config.windowLevel ?? UIWindow.Level.normal
         if #available(iOS 13, *) {
             window.overrideUserInterfaceStyle = config.overrideUserInterfaceStyle
         }
     }
-    
-    func install(becomeKey: Bool) {
-        show(becomeKey: becomeKey)
+
+    func install() {
+        if #available(iOS 13, *) {
+            window?.windowScene = config.windowScene
+            if config.shouldBecomeKeyWindow {
+                previousKeyWindow = UIApplication.shared.keyWindow
+            }
+            show(
+                becomeKey: config.shouldBecomeKeyWindow,
+                frame: config.windowScene?.coordinateSpace.bounds
+            )
+        } else {
+            show(becomeKey: config.shouldBecomeKeyWindow)
+        }
+
     }
 
-    @available(iOS 13, *)
-    func install(becomeKey: Bool, scene: UIWindowScene?) {
-        window?.windowScene = scene
-        if becomeKey {
-            previousKeyWindow = UIApplication.shared.keyWindow
-        }
-        show(becomeKey: becomeKey, frame: scene?.coordinateSpace.bounds)
-    }
-    
     private func show(becomeKey: Bool, frame: CGRect? = nil) {
         guard let window = window else { return }
         window.frame = frame ?? UIScreen.main.bounds
@@ -75,16 +81,14 @@ open class WindowViewController: UIViewController
 
     // MARK: - Variables
 
-    let windowLevel: UIWindow.Level
-
     private var window: UIWindow?
     private weak var previousKeyWindow: UIWindow?
 
-    private let config: SwiftMessages.Config
+    let config: SwiftMessages.Config
 }
 
 extension WindowViewController {
-    static func newInstance(windowLevel: UIWindow.Level?, config: SwiftMessages.Config) -> WindowViewController {
-        return config.windowViewController?(windowLevel, config) ?? WindowViewController(windowLevel: windowLevel, config: config)
+    static func newInstance(config: SwiftMessages.Config) -> WindowViewController {
+        return config.windowViewController?(config) ?? WindowViewController(config: config)
     }
 }